Move Family to its own module to avoid circular imports
