+++ lms/product/product.py
@@ -5,23 +5,7 @@ from enum import Enum
 from typing import Dict, Optional
 
 from lms.product.plugin import PluginConfig, Plugins
-
-
-class Family(str, Enum):
-    """Enum for which product this relates to."""
-
-    BLACKBAUD = "BlackbaudK12"
-    BLACKBOARD = "BlackboardLearn"
-    CANVAS = "canvas"
-    D2L = "desire2learn"
-    MOODLE = "moodle"
-    SAKAI = "sakai"
-    SCHOOLOGY = "schoology"
-    UNKNOWN = "unknown"
-
-    @classmethod
-    def _missing_(cls, _value):
-        return cls.UNKNOWN
+from lms.product.family import Family
 
 
 @dataclass
